requirements.txt
# Dependencias principales
requests==2.28.2
numpy==1.23.4

# Herramientas para análisis de red
scapy==2.4.5

# Manejo de variables de entorno
python-dotenv==1.0.0

# Conectividad BLE
bleak==0.19.1

# Dependencias de desarrollo
pytest==7.4.0

django>=4.0,<5.0
